% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{abc}
\alias{abc}
\title{Approximate Bayesian Computation}
\usage{
abc(data0, fun, epsilon = 1e-06, iter = 1000, summary_fun = c(mean, var),
  distance_fun = l2_norm)
}
\arguments{
\item{data0}{The data}

\item{fun}{The model}

\item{epsilon}{Convergence tolerance}

\item{iter}{Maximum number of iterations}

\item{summary_fun}{Summary statistics of the data and simulated data}

\item{distance_fun}{Distance between two summary statistics}
}
\description{
Approximate Bayesian Computation
}
\examples{
data0 <- rgamma(5000, 5, 2)
generative_fun <- function(alpha, beta) {rgamma(5000, alpha, beta)}
abc(data0, generative_fun)
}
