% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter.R
\name{dispatch_eval}
\alias{dispatch_eval}
\title{Dispatch evaluation function based on keywords}
\usage{
dispatch_eval(cmd, env)
}
\description{
Dispatch evaluation function based on keywords
}
\keyword{internal}
