% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter.R
\name{run_script}
\alias{run_script}
\title{Run script}
\usage{
run_script(str0, env = new.env())
}
\arguments{
\item{str0}{string; the script to be run.}

\item{env}{The environment.}
}
\description{
Run script
}
